% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirror.R
\name{mirror}
\alias{mirror}
\title{Mirror an image}
\usage{
mirror(image_path, direction = "all")
}
\arguments{
\item{image_path}{string: the file path of the image to be mirrored}

\item{direction}{string: dirction of mirroring. Options: 'horizontal', 'vertical', 'all'. Defaults to 'all'}
}
\value{
array/matrix: pixel values of the original image and mirrored image(s)
}
\description{
Mirror takes the path to an image and generates a mirrored version of that image in the horizontal direction, vertical direction, or both.
It returns an array of pixel values for the mirrored image(s).
}
