% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate.R
\name{rotate}
\alias{rotate}
\title{Returns an array of images of length num_images randomly rotated a random degree up to a max_rotation}
\usage{
rotate(image_path, num_images, max_rotation)
}
\arguments{
\item{image_path}{string: The file path of the image to be translated.}

\item{num_images}{integer: The number of rotated images to be returned.}

\item{max_rotation}{integer: The maximum allowable rotation for any of the rotated images. Can be between 1 and 360, inclusive.}
}
\value{
array: pixel values of the rotated images
}
\description{
Rotate takes the path to an image and generates randomly rotated images, the desired number
of times. Each rotated image will not be rotated more than the maximum rotation angle provided.
}
\examples{
rotate("tests/testthat/img/milad.jpg", 5, 180)
rotate("tests/testthat/img/milad.jpg", 10, 320)
rotate("tests/testthat/img/milad.jpg", 15, 350)
rotate("tests/testthat/img/milad.jpg", 20, 355)
}
