% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{translate}
\alias{translate}
\title{Returns an array of images of length num_images randomly translated a random number of pixels up to max_rotation}
\usage{
translate(image_path, num_images, max_translation)
}
\arguments{
\item{image_path}{string: The file path of the image to be translated.}

\item{num_images}{integer: The number of translated images to be returned.}

\item{max_translation}{integer: The maximum distance in pixels that the image can be translated.}
}
\value{
array: pixel values of the translated images
}
\description{
Translate takes the path to and image and generates randomly translated images, the desired number
of times. Each translated image will not be translated more than the maximum distance provided.
The translation can be in any direction.
}
\examples{
\dontrun{
translate("../tests/testthat/img/milad.jpg", 5, 10)
translate("../tests/testthat/img/milad.jpg", 10, 20)
translate("../tests/testthat/img/milad.jpg", 15, 50)
translate("../tests/testthat/img/milad.jpg", 20, 100)
}
}
